unified-devops-nexus/requirements.txt
Click==8.0.3
PyYAML==5.4.1
asyncio==3.4.3